+++ src/extension.ts
@@ -25,6 +25,10 @@ export function activate(context: vscode.ExtensionContext) {
         platformio.installLibrary();
     }));
 
+    context.subscriptions.push(vscode.commands.registerCommand("platformio.openTerminal", () => {
+        platformio.openTerminal();
+    }));
+
     context.subscriptions.push(vscode.window.onDidCloseTerminal((closedTerminal: vscode.Terminal) => {
         platformio.onDidCloseTerminal(closedTerminal);
     }));