+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.4 (2017-04-02)
+* [#3](https://github.com/formulahendry/vscode-platformio/issues/3): Add config entry to set baud rate for Serial Monitor
+* Add 'Open PlatformIO Terminal' icon in Status Bar
+
 ## 0.0.3 (2017-03-25)
 * Add Library Install in Status Bar